jjp-155: rename imp to incoming and exp to outgoing
