+++ app/backend/src/main/java/com/backend/service/user/UserService.java
@@ -1,7 +1,7 @@
 package com.backend.service.user;
 
-import com.backend.entity.dto.exp.RegisterResponse;
-import com.backend.entity.dto.imp.RegisterRequest;
+import com.backend.entity.dto.outgoing.RegisterResponse;
+import com.backend.entity.dto.incoming.RegisterRequest;
 
 public interface UserService {
