+++ app/backend/src/main/java/com/backend/service/user/UserServiceImpl.java
@@ -1,7 +1,7 @@
 package com.backend.service.user;
 
-import com.backend.entity.dto.exp.RegisterResponse;
-import com.backend.entity.dto.imp.RegisterRequest;
+import com.backend.entity.dto.outgoing.RegisterResponse;
+import com.backend.entity.dto.incoming.RegisterRequest;
 import com.backend.entity.user.User;
 import com.backend.entity.user.UserType;
 import com.backend.exception.user.UserFoundException;
