+++ app/backend/src/test/java/com/backend/controller/user/UserControllerTest.java
@@ -1,7 +1,7 @@
 package com.backend.controller.user;
 
-import com.backend.entity.dto.exp.RegisterResponse;
-import com.backend.entity.dto.imp.RegisterRequest;
+import com.backend.entity.dto.outgoing.RegisterResponse;
+import com.backend.entity.dto.incoming.RegisterRequest;
 import com.backend.exception.user.UserFoundException;
 import com.backend.service.user.UserService;
 import org.junit.jupiter.api.Nested;
