+++ app/backend/src/test/java/com/backend/service/user/UserServiceTest.java
@@ -1,12 +1,11 @@
 package com.backend.service.user;
 
-import com.backend.entity.dto.exp.RegisterResponse;
-import com.backend.entity.dto.imp.RegisterRequest;
+import com.backend.entity.dto.outgoing.RegisterResponse;
+import com.backend.entity.dto.incoming.RegisterRequest;
 import com.backend.entity.user.User;
 import com.backend.entity.user.UserType;
 import com.backend.exception.user.UserFoundException;
 import com.backend.repository.user.UserRepository;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
