+++ auth_app/src/main/java/com/blankfactor/auth/entity/dto/incoming/InformRequest.java
@@ -1,4 +1,4 @@
-package com.blankfactor.auth.entity.dto.imp;
+package com.blankfactor.auth.entity.dto.incoming;
 
 import lombok.*;
