+++ auth_app/src/main/java/com/blankfactor/auth/entity/dto/outgoing/LoginResponse.java
@@ -1,4 +1,4 @@
-package com.blankfactor.auth.entity.dto.exp;
+package com.blankfactor.auth.entity.dto.outgoing;
 
 import lombok.Getter;
 import lombok.Setter;
