+++ auth_app/src/main/java/com/blankfactor/auth/entity/dto/outgoing/VerifyResponse.java
@@ -1,4 +1,4 @@
-package com.blankfactor.auth.entity.dto.exp;
+package com.blankfactor.auth.entity.dto.outgoing;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
