+++ auth_app/src/main/java/com/blankfactor/auth/service/AuthService.java
@@ -1,17 +1,17 @@
 package com.blankfactor.auth.service;
 
 import com.blankfactor.auth.entity.Role;
-import com.blankfactor.auth.entity.dto.imp.InformRequest;
-import com.blankfactor.auth.entity.dto.imp.LoginRequest;
+import com.blankfactor.auth.entity.dto.incoming.InformRequest;
+import com.blankfactor.auth.entity.dto.incoming.LoginRequest;
 import com.blankfactor.auth.exception.custom.*;
 import com.blankfactor.auth.exception.custom.code.ExpiredVerificationCodeException;
 import com.blankfactor.auth.exception.custom.code.IncorrectVerificationCodeException;
 import com.blankfactor.auth.exception.custom.user.*;
 import com.blankfactor.auth.entity.User;
-import com.blankfactor.auth.entity.dto.exp.RegisterResponse;
-import com.blankfactor.auth.entity.dto.exp.VerifyResponse;
-import com.blankfactor.auth.entity.dto.imp.RegisterRequest;
-import com.blankfactor.auth.entity.dto.imp.VerifyRequest;
+import com.blankfactor.auth.entity.dto.outgoing.RegisterResponse;
+import com.blankfactor.auth.entity.dto.outgoing.VerifyResponse;
+import com.blankfactor.auth.entity.dto.incoming.RegisterRequest;
+import com.blankfactor.auth.entity.dto.incoming.VerifyRequest;
 import com.blankfactor.auth.repository.UserRepository;
 import jakarta.mail.MessagingException;
 import lombok.RequiredArgsConstructor;
@@ -23,7 +23,6 @@ import org.springframework.http.MediaType;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
