+++ auth_app/src/test/java/com/blankfactor/auth/service/AuthServiceIT.java
@@ -1,10 +1,10 @@
 package com.blankfactor.auth.service;
 
 import com.blankfactor.auth.entity.User;
-import com.blankfactor.auth.entity.dto.exp.RegisterResponse;
-import com.blankfactor.auth.entity.dto.exp.VerifyResponse;
-import com.blankfactor.auth.entity.dto.imp.RegisterRequest;
-import com.blankfactor.auth.entity.dto.imp.VerifyRequest;
+import com.blankfactor.auth.entity.dto.outgoing.RegisterResponse;
+import com.blankfactor.auth.entity.dto.outgoing.VerifyResponse;
+import com.blankfactor.auth.entity.dto.incoming.RegisterRequest;
+import com.blankfactor.auth.entity.dto.incoming.VerifyRequest;
 import com.blankfactor.auth.exception.custom.PasswordsDoNotMatchException;
 import com.blankfactor.auth.exception.custom.code.ExpiredVerificationCodeException;
 import com.blankfactor.auth.exception.custom.code.IncorrectVerificationCodeException;
